Ajuste nos Repository´s

+++ advocacy/src/main/java/github/io/advocacy/repository/ClientRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface ClientRepository extends JpaRepository<ClientEntity, Integer> {
+public interface ClientRepository extends JpaRepository<ClientEntity, Long> {
 
 }
