+++ advocacy/src/main/java/github/io/advocacy/repository/LawyerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LawyerRepository extends JpaRepository<LawyerRepository, Integer> {
+public interface LawyerRepository extends JpaRepository<LawyerRepository, Long> {
 
 }
